finally

+++ src/io/puts.c
@@ -6,11 +6,11 @@
 #include <string.h>
 int puts(const char *str)
 {
-    size_t nr = strlen(str);
-    size_t apel = syscall(1, 1, str, nr);
+    int nr = strlen(str);
+    int apel = syscall(1, 1, str, nr);
     if (apel == -1)
         return -1;
-    size_t apel2 = syscall(1, 1, "\n", 1);
+    int apel2 = syscall(1, 1, "\n", 1);
     if (apel2 == -1)
         return -1;
     return 0;
